refactored utility functions into src folder
