+++ templates/about.html
@@ -29,7 +29,7 @@
         Please always contact the foodbank in advance of making referrals or travelling for help.
     </p>
     <p>
-        Our underlying manually maintained dataset can be downloaded in full from <a href={{ url_for("static", filename="foodbanks.csv") }}>here</a>.
+        The website's underlying dataset, which we maintain manually, can be downloaded in full from <a href={{ url_for("static", filename="foodbanks.csv") }}>here</a>.
     </p>
 </div>
